# Conflicts:
#	Bundles/Sales/src/Spryker/Zed/Sales/Business/OrderItem/SalesOrderItemReader.php
#	Bundles/ShipmentGui/src/Spryker/Zed/ShipmentGui/Dependency/Facade/ShipmentGuiToSalesFacadeBridge.php
#	Bundles/ShipmentGui/src/Spryker/Zed/ShipmentGui/Dependency/Facade/ShipmentGuiToSalesFacadeInterface.php

+++ Bundles/ProductPackagingUnit/src/Spryker/Zed/ProductPackagingUnit/Persistence/ProductPackagingUnitRepositoryInterface.php
@@ -107,7 +107,7 @@ interface ProductPackagingUnitRepositoryInterface
      * @param string $sku
      * @param string[] $reservedStateNames
      *
-     * @return int
+     * @return float
      */
-    public function sumLeadProductAmountForAllSalesOrderItemsBySku(string $sku, array $reservedStateNames): int;
+    public function sumLeadProductAmountForAllSalesOrderItemsBySku(string $sku, array $reservedStateNames): float;
 }
